add(authenticator):

+++ controllers/updateDirectorController.js
@@ -5,6 +5,7 @@
 
 const mysql = require('mysql2')
 require('dotenv').config()
+const auth = require('../validators/authenticator')
 const controller = {}
 
 /**
@@ -16,41 +17,45 @@ const controller = {}
  * @param {object} res the Express response.
  */
 controller.update = async (req, res) => {
-  try {
-    /**
-     * DB connection.
-     */
-    const db = mysql.createPool({
-      connectionLimit: 100,
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_NAME
-    })
-
-    /**
-     * Exporting the DB connection.
-     *
-     * @param {object} req the Express request.
-     * @param {object} res the Express response.
-     */
-
-    db.getConnection((error, connection) => {
-      if (error) {
-        console.log(error)
-        process.exit(1)
-      } else {
-        console.log('MySQL is connected. Connection ID: ' + connection.threadId)
-      }
-      connection.query('SELECT * FROM directors_table WHERE id = ?', [req.params.id], (err, rows) => {
-        connection.release()
-        if (!err) {
-          res.render('update/directors', { rows, title: 'Update Director' })
-        }
+  if (auth.checkAuthenticated(req)) {
+    try {
+      /**
+       * DB connection.
+       */
+      const db = mysql.createPool({
+        connectionLimit: 100,
+        host: process.env.DB_HOST,
+        user: process.env.DB_USER,
+        password: process.env.DB_PASSWORD,
+        database: process.env.DB_NAME
       })
-    })
-  } catch (error) {
-    console.log(error)
+
+      /**
+       * Exporting the DB connection.
+       *
+       * @param {object} req the Express request.
+       * @param {object} res the Express response.
+       */
+
+      db.getConnection((error, connection) => {
+        if (error) {
+          console.log(error)
+          process.exit(1)
+        } else {
+          console.log('MySQL is connected. Connection ID: ' + connection.threadId)
+        }
+        connection.query('SELECT * FROM directors_table WHERE id = ?', [req.params.id], (err, rows) => {
+          connection.release()
+          if (!err) {
+            res.render('update/directors', { rows, title: 'Update Director' })
+          }
+        })
+      })
+    } catch (error) {
+      console.log(error)
+    }
+  } else {
+    return res.redirect('/login')
   }
 }
 
@@ -64,41 +69,45 @@ controller.update = async (req, res) => {
 controller.updateDirector = async (req, res) => {
   const { firstName, lastName, age, origin, id } = req.body
   const fullName = firstName + ' ' + lastName
-  try {
-  /**
-   * DB connection.
-   */
-    const db = mysql.createPool({
-      connectionLimit: 100,
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_NAME
-    })
-
-    /**
-     * Exporting the DB connection.
-     *
-     * @param {object} req the Express request.
-     * @param {object} res the Express response.
-     */
-
-    db.getConnection((error, connection) => {
-      if (error) {
-        console.log(error)
-        process.exit(1)
-      } else {
-        console.log('MySQL is connected. Connection ID: ' + connection.threadId)
-      }
-      connection.query('UPDATE directors_table d, movies_table m, series_table s SET d.firstName = ?, d.lastName = ?, d.age = ?, d.fullName = ?, d.origin = ?, m.director = ?, s.director = ? WHERE d.id = ? AND m.directorID = ? AND s.directorID = ?', [firstName, lastName, age, fullName, origin, fullName, fullName, id, id, id], (err, rows) => {
-        connection.release()
-        if (!err) {
-          res.redirect('/directors')
-        }
+  if (auth.checkAuthenticated(req)) {
+    try {
+      /**
+       * DB connection.
+       */
+      const db = mysql.createPool({
+        connectionLimit: 100,
+        host: process.env.DB_HOST,
+        user: process.env.DB_USER,
+        password: process.env.DB_PASSWORD,
+        database: process.env.DB_NAME
       })
-    })
-  } catch (error) {
-    console.log(error)
+
+      /**
+       * Exporting the DB connection.
+       *
+       * @param {object} req the Express request.
+       * @param {object} res the Express response.
+       */
+
+      db.getConnection((error, connection) => {
+        if (error) {
+          console.log(error)
+          process.exit(1)
+        } else {
+          console.log('MySQL is connected. Connection ID: ' + connection.threadId)
+        }
+        connection.query('UPDATE directors_table d, movies_table m, series_table s SET d.firstName = ?, d.lastName = ?, d.age = ?, d.fullName = ?, d.origin = ?, m.director = ?, s.director = ? WHERE d.id = ? AND m.directorID = ? AND s.directorID = ?', [firstName, lastName, age, fullName, origin, fullName, fullName, id, id, id], (err, rows) => {
+          connection.release()
+          if (!err) {
+            res.redirect('/directors')
+          }
+        })
+      })
+    } catch (error) {
+      console.log(error)
+    }
+  } else {
+    return res.redirect('/login')
   }
 }
 
@@ -110,41 +119,45 @@ controller.updateDirector = async (req, res) => {
  * @param {object} res the Express response.
  */
 controller.delete = async (req, res) => {
-  try {
-  /**
-   * DB connection.
-   */
-    const db = mysql.createPool({
-      connectionLimit: 100,
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_NAME
-    })
-
-    /**
-     * Exporting the DB connection.
-     *
-     * @param {object} req the Express request.
-     * @param {object} res the Express response.
-     */
-
-    db.getConnection((error, connection) => {
-      if (error) {
-        console.log(error)
-        process.exit(1)
-      } else {
-        console.log('MySQL is connected. Connection ID: ' + connection.threadId)
-      }
-      connection.query('DELETE FROM directors_table WHERE id = ?', [req.params.id], (err, rows) => {
-        connection.release()
-        if (!err) {
-          res.redirect('/directors')
-        }
+  if (auth.checkAuthenticated(req)) {
+    try {
+      /**
+       * DB connection.
+       */
+      const db = mysql.createPool({
+        connectionLimit: 100,
+        host: process.env.DB_HOST,
+        user: process.env.DB_USER,
+        password: process.env.DB_PASSWORD,
+        database: process.env.DB_NAME
       })
-    })
-  } catch (error) {
-    console.log(error)
+
+      /**
+       * Exporting the DB connection.
+       *
+       * @param {object} req the Express request.
+       * @param {object} res the Express response.
+       */
+
+      db.getConnection((error, connection) => {
+        if (error) {
+          console.log(error)
+          process.exit(1)
+        } else {
+          console.log('MySQL is connected. Connection ID: ' + connection.threadId)
+        }
+        connection.query('DELETE FROM directors_table WHERE id = ?', [req.params.id], (err, rows) => {
+          connection.release()
+          if (!err) {
+            res.redirect('/directors')
+          }
+        })
+      })
+    } catch (error) {
+      console.log(error)
+    }
+  } else {
+    return res.redirect('/login')
   }
 }
